requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
google-generativeai==0.8.5
httpx==0.28.1
tenacity==8.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
drf-spectacular==0.27.2
Pillow==10.3.0
python-dotenv==1.0.1
twilio==9.2.3
channels==4.1.0       
channels-redis==4.2.0
pdfplumber==0.11.0
openpyxl==3.1.2
pandas==2.2.2
chardet==5.2.0
google-genai==1.39.1